pyproject.toml
[tool.poetry]
name = "nxcmdr"
version = "0.1.0"
description = ""
authors = ["xyder <xyder@dsider.org>"]

[tool.poetry.dependencies]
python = "^3.8"
plumbum = "^1.6.9"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
maturin = "^0.8.3"
pylint = "^2.6.0"
ipdb = "^0.13.4"

[tool.poetry.scripts]
nxc = 'nxcmdr.__main__:main'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

